requirements.txt
asgiref==3.8.1
Django==5.1.6
django-filter==25.1
djangorestframework==3.15.2
Markdown==3.7
psycopg2-binary==2.9.10
sqlparse==0.5.3
